# swagger.yaml
swagger: "2.0"
info:
  description: "here is my Order service api documentation."
  version: "1.0.0"
  title: "Swagger Order"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8083"
basePath: "/"
schemes:
- "http"
- "https"
paths:   
...
  /orders/update-order/{id}:  
    put:
      tags:
      - "order"
      summary: "Update an existing order"
      description: ""
      operationId: "updateOrder"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Pet"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
        "405":
          description: "Validation exception"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
...